askracha/bot/requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0